Consolidate synth package
Moved all code into synth.py
Renamed Note to SynthNote
Moved test yaml into test/data folder
Moved example file into test_integration
