--- plaintext_daw/models/synth/.gitignore
@@ -1,1 +1,0 @@
-*.wav
